fix(l2): cast AvtpFrameHeader::sequence_num (uint16_t) to uint8_t for FrameHandler calls in tests
Fixes an MSVC /W4 /WX narrowing-conversion error (C2220/C4244) at two
new test call sites (test_l2.cpp:525,572) that call FrameHandler's
uint8_t sequence_num parameter directly with hdr.sequence_num
(uint16_t) -- not caught by clang/gcc's warning sets on the other CI
platforms. Test-only, no production code affected.

+++ tests/test_l2.cpp
@@ -522,7 +522,7 @@ TEST_CASE("FrameHandler wired to mock::Server::dispatch_frame() dispatches every
     std::vector<uint8_t> raw(wire.begin() + static_cast<long>(hdr.acf_offset), wire.end());
 
     std::vector<FrameMemberResult> results;
-    size_t n = fh(/*client=*/0, hdr.stream_id, hdr.sequence_num, raw, results);
+    size_t n = fh(/*client=*/0, hdr.stream_id, static_cast<uint8_t>(hdr.sequence_num), raw, results);
     REQUIRE(n == 2);
     REQUIRE(results.size() == 2);
     REQUIRE(results[0].byte_bus_id == mock::kGpioByteBusId);
@@ -569,7 +569,7 @@ TEST_CASE("FrameHandler wired to mock::Server::dispatch_frame_e2e() is reachable
     std::vector<uint8_t> raw(wire.begin() + static_cast<long>(hdr.acf_offset), wire.end());
 
     std::vector<FrameMemberResult> results;
-    size_t n = fh(/*client=*/0, hdr.stream_id, hdr.sequence_num, raw, results);
+    size_t n = fh(/*client=*/0, hdr.stream_id, static_cast<uint8_t>(hdr.sequence_num), raw, results);
     REQUIRE(n == 1);
     REQUIRE(results.size() == 1);
     REQUIRE(results[0].response.rsp);
